Changing the Health from int to decimal

+++ RandomHeroGenerator.Host/Models/BattleRound.cs
@@ -6,7 +6,7 @@
         public string? DefenderType { get; set; }
         public int Attacker { get; set; }
         public int Defender { get; set; }
-        public int AttackerHealthChange { get; set; }
-        public int DefenderHealthChange { get; set; }
+        public decimal AttackerHealthChange { get; set; }
+        public decimal DefenderHealthChange { get; set; }
     }
 }
